+++ src/schemas/interfaces/calendar-source.ts
@@ -5,14 +5,17 @@ import type { CalendarSourceType } from '../enums/calendar-source-type';
  */
 export interface CalendarSource {
   /**
+   * @platform iOS
    * @since 7.1.0
    */
   type: CalendarSourceType;
   /**
+   * @platform iOS
    * @since 7.1.0
    */
   id: string;
   /**
+   * @platform iOS
    * @since 7.1.0
    */
   title: string;